Fix floating point exception.
With gfortran 9 and later, the code might generate an exception
with the message

Program received signal SIGFPE: Floating-point exception - erroneous arithmetic operation.

and point to the code at dde_solver_m.f90:6274.  MY_TRIM_FREQUENCY
can be 0, and there is an expression that divides by MY_TRIM_FREQUENCY
that apparently wasn't evaluated in gfortran 8, but is evaluated in newer
versions of gfortran.

Thanks to Arjen Markus (github user @arjenmarkus) for finding the
bug and suggesting the fix.